requirements.txt
streamlit>=1.33
google-generativeai>=0.7
chromadb==0.5.5
posthog==2.4.2          # <- silences the capture() mismatch
networkx>=3.2
python-dotenv>=1.0
pydantic>=2.7
typing_extensions>=4.9
pysqlite3-binary; platform_system == "Linux"
fastapi>=0.110
uvicorn>=0.29
